+++ common-http/src/main/java/cn/featherfly/common/http/HttpRequest.java
@@ -113,37 +113,6 @@ public interface HttpRequest {
     /**
      * 发送请求（请求体作为参数），异步网络请求，使用Future转换为同步方法调用方式.
      *
-     * @param <T>                   返回内容转换的对象类型
-     * @param method                httpmethod
-     * @param url                   url
-     * @param params                请求参数
-     * @param headers               请求头
-     * @param responseType          返回类型
-     * @param errorListener         错误监听器
-     * @param requestTimeoutSeconds 请求超时时间
-     * @return 返回内容转换后的对象
-     */
-    <T> T send(HttpMethod method, String url, Map<String, String> params, Map<String, String> headers,
-            Class<T> responseType, ErrorListener errorListener, long requestTimeoutSeconds);
-
-    /**
-     * 发送请求（请求体作为参数），异步网络请求，使用Future转换为同步方法调用方式.
-     *
-     * @param <T>                   返回内容转换的对象类型
-     * @param method                httpmethod
-     * @param url                   url
-     * @param params                请求参数
-     * @param responseType          返回类型
-     * @param errorListener         错误监听器
-     * @param requestTimeoutSeconds 请求超时时间
-     * @return 返回内容转换后的对象
-     */
-    <T> T send(HttpMethod method, String url, Map<String, String> params, Class<T> responseType,
-            ErrorListener errorListener, long requestTimeoutSeconds);
-
-    /**
-     * 发送请求（请求体作为参数），异步网络请求，使用Future转换为同步方法调用方式.
-     *
      * @param <T>           返回内容转换的对象类型
      * @param method        httpmethod
      * @param url           url