+++ common-http/src/test/java/cn/featherfly/common/http/BaiduMapTest.java
@@ -4,6 +4,8 @@ package cn.featherfly.common.http;
 import java.util.HashMap;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import org.testng.annotations.Test;
+
 import cn.featherfly.common.http.po.ReverseGeocoding;
 
 /**
@@ -13,12 +15,14 @@ import cn.featherfly.common.http.po.ReverseGeocoding;
  */
 public class BaiduMapTest {
 
-    public static void main(String[] args) {
+    @Test
+    public void testBrower() {
         OkHttpRequest request = new OkHttpRequest(new HttpRequestConfig());
 
-        ReverseGeocoding reverseGeocoding = request.send(HttpMethod.GET,
-                "http://api.map.baidu.com/reverse_geocoding/v3/?ak=0Zj0WEeN8dTr8XMMAFal5LPE8k5D53tR&output=json&coordtype=wgs84ll&location=31.225696563611,121.49884033194",
-                new HashMap<>(), ReverseGeocoding.class, (ErrorListener) error -> {
+        String url = "http://api.map.baidu.com/reverse_geocoding/v3/?ak=wwhuWsqLYCMuFPE8fNVPk4iUNzYCIFAk&output=json&coordtype=wgs84ll&location=30.664416,104.072294";
+
+        ReverseGeocoding reverseGeocoding = request.send(HttpMethod.GET, url, new HashMap<>(), ReverseGeocoding.class,
+                (ErrorListener) error -> {
                     System.out.println(error.getMessage());
                 });
 
@@ -26,15 +30,13 @@ public class BaiduMapTest {
 
         AtomicBoolean executed = new AtomicBoolean(false);
 
-        request.send(HttpMethod.GET,
-                "http://api.map.baidu.com/reverse_geocoding/v3/?ak=0Zj0WEeN8dTr8XMMAFal5LPE8k5D53tR&output=json&coordtype=wgs84ll&location=31.225696563611,121.49884033194",
-                new HashMap<>(), ReverseGeocoding.class).completion(r -> {
-                    System.out.println(r);
-                    executed.set(true);
-                }).error(e -> {
-                    System.out.println(e.getMessage());
-                    executed.set(true);
-                });
+        request.send(HttpMethod.GET, url, new HashMap<>(), ReverseGeocoding.class).completion(r -> {
+            System.out.println(r);
+            executed.set(true);
+        }).error(e -> {
+            System.out.println(e.getMessage());
+            executed.set(true);
+        });
 
         while (!executed.get()) {
 
@@ -43,6 +45,43 @@ public class BaiduMapTest {
         request.shutdown();
 
         System.out.println("end");
+
+    }
+
+    @Test
+    public void testServer() {
+        OkHttpRequest request = new OkHttpRequest(new HttpRequestConfig());
+
+        String url = "http://api.map.baidu.com/reverse_geocoding/v3/?ak=0Zj0WEeN8dTr8XMMAFal5LPE8k5D53tR&output=json&coordtype=wgs84ll&location=31.225696563611,121.49884033194";
+
+        ReverseGeocoding reverseGeocoding = request.send(HttpMethod.GET, url, new HashMap<>(), ReverseGeocoding.class,
+                (ErrorListener) error -> {
+                    System.out.println(error.getMessage());
+                });
+
+        System.out.println(reverseGeocoding);
+
+        AtomicBoolean executed = new AtomicBoolean(false);
+
+        request.send(HttpMethod.GET, url, new HashMap<>(), ReverseGeocoding.class).completion(r -> {
+            System.out.println(r);
+            executed.set(true);
+        }).error(e -> {
+            System.out.println(e.getMessage());
+            executed.set(true);
+        });
+
+        while (!executed.get()) {
+
+        }
+
+        request.shutdown();
+
+        System.out.println("end");
+    }
+
+    public static void main(String[] args) {
+
         //        Http.download(
         //                "http://api.map.baidu.com/reverse_geocoding/v3/?ak=0Zj0WEeN8dTr8XMMAFal5LPE8k5D53tR&output=json&coordtype=wgs84ll&location=31.225696563611,121.49884033194",
         //                new File("reverse_geocoding.json"));